# Record & console log performance metrics