--- a/pxtlib/analytics.ts
+++ b/pxtlib/analytics.ts
@@ -4,13 +4,16 @@ namespace pxt {
     // These functions are defined in docfiles/pxtweb/cookieCompliance.ts
     export declare function aiTrackEvent(id: string, data?: any, measures?: any): void;
     export declare function aiTrackException(err: any, kind: string, props: any): void;
+}
 
-    export declare namespace perf {
-        export function reset(): void;
-        export function splitMs(): number;
-        export function splitStr(): string;
-        export function logSplit(msg: string): void;
-    }
+namespace pxt.perf {
+    // These functions are defined in docfiles/pxtweb/cookieCompliance.ts
+    export declare function reset(): void;
+    export declare function splitMs(): number;
+    export declare function splitStr(): string;
+    export declare function recordMilestone(msg: string, time?: number): void;
+    export declare function measureStart(name: string): void;
+    export declare function measureEnd(name: string): void;
 }
 
 namespace pxt.analytics {