--- a/pxtlib/analytics.ts
+++ b/pxtlib/analytics.ts
@@ -8,9 +8,7 @@ namespace pxt {
 
 namespace pxt.perf {
     // These functions are defined in docfiles/pxtweb/cookieCompliance.ts
-    export declare function reset(): void;
-    export declare function splitMs(): number;
-    export declare function splitStr(): string;
+    export declare function report(): void;
     export declare function recordMilestone(msg: string, time?: number): void;
     export declare function measureStart(name: string): void;
     export declare function measureEnd(name: string): void;