--- a/pxtlib/workeriface.ts
+++ b/pxtlib/workeriface.ts
@@ -61,7 +61,9 @@ namespace pxt.worker {
         let worker = new Worker(workerFile)
         let iface = wrap(v => worker.postMessage(v))
         worker.onmessage = ev => {
+            pxt.perf.measureStart("webworker recvHandler")
             iface.recvHandler(ev.data)
+            pxt.perf.measureEnd("webworker recvHandler")
         }
         return iface
     }