@pelikhan 2019-11-21
A format that would be importeable in Excel would be helpful to do further analysis.