@pelikhan 2019-11-21
Can we log the report to analytics as well?